Reuse EntityList and modify some wording
