+++ Source/Plugins/Core/com.equella.core/js/tsrc/cloudprovider/CloudProviderEntity.ts
@@ -1,6 +1,6 @@
 export interface CloudProviderEntity {
   id: string;
   name: string;
-  description: string;
-  iconUrl: string;
+  description?: string;
+  iconUrl?: string;
 }
