+++ Source/Plugins/Core/com.equella.core/js/tsrc/cloudprovider/CloudProviderListPage.tsx
@@ -1,18 +1,32 @@
 import { Bridge } from "../api/bridge";
 import * as React from "react";
-import { Fab, withStyles, WithStyles } from "@material-ui/core";
-import AddIcon from "@material-ui/icons/Add";
-import CloudProviderList from "./CloudProviderList";
-import { CloudProviderEntity } from "./CloudProviderEntity";
 import {
-  getCloudProviderListPageStyle,
-  getCloudProviders,
-  langStrings
-} from "./CloudProviderModule";
+  Avatar,
+  createStyles,
+  IconButton,
+  ListItem,
+  ListItemAvatar,
+  ListItemSecondaryAction,
+  ListItemText,
+  Theme,
+  Typography,
+  withStyles,
+  WithStyles
+} from "@material-ui/core";
+import DeleteIcon from "@material-ui/icons/Delete";
+import { CloudProviderEntity } from "./CloudProviderEntity";
+import { getCloudProviders, langStrings } from "./CloudProviderModule";
 import { AxiosError } from "axios";
 import { ErrorResponse, generateFromError } from "../api/errors";
+import EntityList from "../components/EntityList";
+import { formatSize } from "../util/langstrings";
 
-const styles = getCloudProviderListPageStyle();
+const styles = (theme: Theme) =>
+  createStyles({
+    searchResultContent: {
+      marginTop: theme.spacing.unit
+    }
+  });
 
 interface CloudProviderListPageProps extends WithStyles<typeof styles> {
   bridge: Bridge;
@@ -51,12 +65,53 @@ class CloudProviderListPage extends React.Component<
   render() {
     const { Template } = this.props.bridge;
     const { cloudProviders, error } = this.state;
+    const createLink = { href: "", onClick: () => {} };
     return (
       <Template title={langStrings.title} errorResponse={error}>
-        <CloudProviderList cloudProviders={cloudProviders} />
-        <Fab className={this.props.classes.fab} color="secondary">
-          <AddIcon />
-        </Fab>
+        <EntityList
+          resultsText={formatSize(
+            cloudProviders.length,
+            langStrings.cloudProviderAvailable
+          )}
+          progress={false}
+          createLink={createLink}
+        >
+          {cloudProviders.map(cloudProvider => {
+            let primaryText = (
+              <Typography color="primary" variant="subtitle1">
+                {cloudProvider.name}
+              </Typography>
+            );
+            let secondaryText = (
+              <Typography
+                variant="body1"
+                className={this.props.classes.searchResultContent}
+              >
+                {cloudProvider.description}
+              </Typography>
+            );
+            return (
+              <ListItem button divider key={cloudProvider.id}>
+                <ListItemAvatar>
+                  <Avatar
+                    alt={cloudProvider.description}
+                    src={cloudProvider.iconUrl}
+                  />
+                </ListItemAvatar>
+                <ListItemText
+                  disableTypography={true}
+                  primary={primaryText}
+                  secondary={secondaryText}
+                />
+                <ListItemSecondaryAction>
+                  <IconButton>
+                    <DeleteIcon />
+                  </IconButton>
+                </ListItemSecondaryAction>
+              </ListItem>
+            );
+          })}
+        </EntityList>
       </Template>
     );
   }
