+++ Source/Plugins/Core/com.equella.core/resources/lang/i18n-resource-centre.properties
@@ -738,7 +738,7 @@ cloneonly.title=Clone item into a collection
 cloneormove.title=Clone or Move item into a collection
 close.confirm=Are you sure you wish to exit the editor?  Any unsaved changes will be lost.
 closebulk=Close
-cloudprovider.settings.title=Cloud Provider
+cloudprovider.settings.title=Cloud providers
 cloudprovider.settings.description=Create and delete cloud providers
 cloud.settings.title=Cloud
 cloud.share.search.email.intro=openEQUELLA Notification\n\nHello,\n\nYou are receiving this email because {0} has shared these cloud search results with you.\n\n
